requirements.txt
flask==3.0.0
flask-cors==4.0.0
spacy==3.7.2
openai==2.0.0
python-dotenv==1.0.0

# Run after install:
# python -m spacy download en_core_web_sm


